[Add]Full test and clean code
